Removed BaseClippingTransform class for the sake of clarity of code
